+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 ## v0.9.0-rc16 [unreleased]
 
+### Features
+- [#2058](https://github.com/influxdb/influxdb/pull/2058): Track number of queries executed in stats.
+- [#2059](https://github.com/influxdb/influxdb/pull/2059): Retention policies sorted by name on return to client.
+- [#2061](https://github.com/influxdb/influxdb/pull/2061): Implement SHOW DIAGNOSTICS.
+- [#2064](https://github.com/influxdb/influxdb/pull/2064): Allow init.d script to return influxd version.
+
 ### Bugfixes
-- [#2037](https://github.com/influxdb/influxdb/pull/2037): Don't check 'configExists' at Run() level
-- [#2039](https://github.com/influxdb/influxdb/pull/2039): Don't panic if getting current use fails
-- [#2034](https://github.com/influxdb/influxdb/pull/2034): Group by should require an aggregate.
-- [#2040](https://github.com/influxdb/influxdb/pull/2040): Add missing top-level help for config command
+- [#2037](https://github.com/influxdb/influxdb/pull/2037): Don't check 'configExists' at Run() level.
+- [#2039](https://github.com/influxdb/influxdb/pull/2039): Don't panic if getting current user fails.
+- [#2034](https://github.com/influxdb/influxdb/pull/2034): GROUP BY should require an aggregate.
+- [#2040](https://github.com/influxdb/influxdb/pull/2040): Add missing top-level help for config command.
+- [#2057](https://github.com/influxdb/influxdb/pull/2057): Move racy "in order" test to integration test suite.
+- [#2060](https://github.com/influxdb/influxdb/pull/2060): Reload server shard map on restart.
 
 ## v0.9.0-rc15 [2015-03-19]